added int

+++ all.py
@@ -16,8 +16,8 @@ Press Ctrl+C to exit!
 ##  --  Scrollphat Setup
 DISPLAY_BAR = True
 while True:
-    if light.light() < 1000:
-        BRIGHTNESS = light.light()/1000
+    if int(light.light()) < 1000:
+        BRIGHTNESS = int(int(light.light())/1000)
     else:
         BRIGHTNESS = 1
 while True:
